requirements.txt
langchain_text_splitters
sentence_transformers
pymongo
python-dotenv